Windows: Consolidate MSVC and MinGW crash handlers
This changes MinGW crash handling to use `SetUnhandledExceptionFilter`,
which handles more crash conditions than using signal handlers. Stack
walking is now done using `StackWalk64`, same as MSVC builds. This makes
stack walking more robust as it can walk through stack frames of system
DLL functions. It still uses `libbacktrace` to symbolize stack frames
within the Godot EXE module, which handles DWARF symbols.
