+++ platform/windows/SCsub
@@ -23,12 +23,8 @@ common_win = [
     "gl_manager_windows_angle.cpp",
     "wgl_detect_version.cpp",
     "rendering_context_driver_vulkan_windows.cpp",
+    "crash_handler_windows_seh.cpp",
 ]
-
-if env.msvc:
-    common_win += ["crash_handler_windows_seh.cpp"]
-else:
-    common_win += ["crash_handler_windows_signal.cpp"]
 
 common_win_wrap = [
     "console_wrapper_windows.cpp",
