+++ platform/windows/crash_handler_windows.h
@@ -34,12 +34,11 @@
 #define WIN32_LEAN_AND_MEAN
 #include <windows.h>
 
-// Crash handler exception only enabled with MSVC
 #if defined(DEBUG_ENABLED)
 #define CRASH_HANDLER_EXCEPTION 1
 
 #ifdef _MSC_VER
-extern DWORD CrashHandlerException(EXCEPTION_POINTERS *ep);
+extern LONG CrashHandlerException(EXCEPTION_POINTERS *ep);
 #endif
 
 #endif
